discord-bot/requirements.txt
discord-py>=2.6.4
anthropic>=0.75.0
pillow>=12.1.0
python-dotenv>=1.2.1
httpx>=0.28.1